# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VideoTranscodeInputBucket = typeof VideoTranscodeInputBucket[keyof typeof VideoTranscodeInputBucket];


export const VideoTranscodeInputBucket = {
  videos: 'videos',
  reels: 'reels',
  moments: 'moments',
} as const;

export interface VideoTranscodeInput {
  bucket: VideoTranscodeInputBucket;
  /**
     * @minLength 3
     * @maxLength 512
     */
  path: string;
}

export interface VideoTranscodeResponse {
  path: string;
  contentType: 'video/mp4';
  faststart: boolean;
}

export interface PostPinInput {
  pinned: boolean;
}

export interface PostPinResponse {
  postId: string;
  pinned: boolean;
}

export interface ApiError {
  error: string;
}