# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: media
    description: Media processing operations
  - name: posts
    description: Post operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /media/transcode:
    post:
      operationId: transcodeVideo
      tags: [media]
      summary: Prepare uploaded video for fast-start playback
      security:
        - SupabaseBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoTranscodeInput"
      responses:
        "200":
          description: Processed MP4 storage object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTranscodeResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Object is not owned by the caller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          description: Video exceeds the processing limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Video processing capacity is busy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Storage or FFmpeg processing failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /posts/{postId}/pin:
    patch:
      operationId: setPostPin
      tags: [posts]
      summary: Set whether an owned post is pinned to its owner's profile grid
      security:
        - SupabaseBearer: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostPinInput"
      responses:
        "200":
          description: Saved pin state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostPinResponse"
        "400":
          description: Invalid post ID or pin state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Post not found or not owned by the caller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: The caller has reached the three-post pin limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Supabase could not complete the pin update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Pinning is unavailable because the posts.pinned column is missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  securitySchemes:
    SupabaseBearer:
      type: http
      scheme: bearer
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    VideoTranscodeInput:
      type: object
      additionalProperties: false
      properties:
        bucket:
          type: string
          enum: [videos, reels, moments]
        path:
          type: string
          minLength: 3
          maxLength: 512
      required:
        - bucket
        - path
    VideoTranscodeResponse:
      type: object
      additionalProperties: false
      properties:
        path:
          type: string
        contentType:
          type: string
          const: video/mp4
        faststart:
          type: boolean
      required:
        - path
        - contentType
        - faststart
    PostPinInput:
      type: object
      additionalProperties: false
      properties:
        pinned:
          type: boolean
      required:
        - pinned
    PostPinResponse:
      type: object
      additionalProperties: false
      properties:
        postId:
          type: string
        pinned:
          type: boolean
      required:
        - postId
        - pinned
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error